client: set position and size for squares and board
Each square is 50x50 (for now; subject to change with VideoMode). Thus
the board dimensions are 400x400. The board should be centered (for now,
until other objects such as chat window and clock are added).

For the board to be centered, its diagonals must intersect at the center
of the window. Thus the board's upper-right corner should be:
	x = centerX - (sqrt(400 * 400 * 2)) / 2
	y = centerY - (sqrt(400 * 400 * 2)) / 2
Basically find the length of the diagonal and divide it by 2. Then
subtract the quotient from the center vertices.