client: draw the board on the window and define Square::setColor()
draw() methods are implemented in both Square and Board classes. The
board just calls each Square's draw().

Also, though Square::setColor() had been declared earlier, I had
forgotten to define it. So that too is done.

In addition, main() function is suitably altered to show the Board
instead of the lone Square that was shown earlier.