client/square: add sf::Vector2f member to store position
Earlier, position was set during the Board's construction using
sf::RectangleShape::setPosition(). But this complicates the movement of
pieces later on. By storing the position during the construction of the
board, all movements later on are simply a matter of calling the
sf::Sprite::move() method with destination Square's position as the
argument.