client: create game class and get indices of clicked square
The game class takes over most of main()'s functions (for now, and maybe
ever).

When a click event is detected, its coordinates are used to determine
the square that was clicked (or if the click was outside the board).
This is done by subtracting the click coordinates with the board's
starting coordinates and dividing the difference by the size of the
square. In addition, for getting the rank, the quotient is subtracted
from 7, since the starting position is actually the top-left. So a
quotient of 6 would imply that a square on rank 1 was clicked.