client/board: implement highlighting of valid move squares
When a piece is clicked, all squares that it can legally move to are now
highlighted.

A std::vector<Square*> Board member is used for this purpose.
Piece::moves() returns legal moves to this member, which are then
highlighted.

New methods -- highlightSquares() and dehighlightSquares() -- are
introduced to abstract away the looping from handleMouseEvt().
Dehighlight is also now modified, hereby requiring the pointer to the
square that needs to be dehighlighted.