client/piece: fix move generation for pawns
The problems were as follows:
1. (rank, file) instead of the correct (file, rank) was being used for
   indexing the board.
2. there were no checks to ensure that the dimensions being generated
   aren't out-of-bounds, causing segmentation faults

These problems are fixed by this commit. A new isMoveValid() method is
introduced to do the checks for diagonal (attack) moves. There is no
reason to not move the line movements checks there as well, but it will
be done later when constraints related to the King's safety are
implemented.