client/pieces: fix straight line movement generation
The problem was that the loop began from the current (i.e. the piece's)
file/rank, and was constrained by the square's emptiness. This resulted
in the loop never executing because the current square was never empty,
owing to the piece's presence.

This behaviour affected both castles and queens.

So loop variable's initialization should be 1 before the file or rank,
for generating leftward and moves respectively, and 1 more for
right-ward and forward.