client/pieces: implement methods for updating properties after moving
For Knight, Bishop and Queen, this property is just m_square. For Pawn,
Castle and King, there is a member variable to be updated.

Piece class defines a (non-pure) virtual method called "move()". The
method is just to standardize the API for the Board class. Its original
definition just calls setSquare(). Its overriding defintions in Pawn,
Castle and King calls setSquare() and also updates other member
variables.

This extra step in Knight, Bishop and Queen is necessary to provide a
standard interface for all pieces.