client/board: simplify dehighlight() method
The colour of a square is stored in the square object and there already
exists a getter method for it, which I had forgotten. The recomputation
of square's original colour is unnecessary. Even the dehighlight()
method is probably unneccessary now, but it remains for the time-being
until this branch is merged.