client/board: generate pieces at beginning and after relevant move
Earlier, moves for each piece were generated only when the piece was
clicked. It is better to generate moves for all pieces during their
initialization. After that, moves for each piece are generated only when
one of the squares they influenced had a piece moved *from* it. (*to* it
is still to be implemented)