client: add file for helper functions
This file would contain functions that are too generic to be exclusive
members of any one class. A lot of common operations would be abstracted
behind functions written in this file.

For now, it contains only one function: to check if the given file and
rank exist on the board.