client/pieces: implement notifying square of Knight's influence
The knight can only move to those squares that are either empty or are
occupied by the opponent's pieces. But it needs to notify all those
squares of its influence on them (regardless of occupational status)
that fall within its reach. Thus the bounds need to be tested for all
possible moves from its positions.

The changes brought in make it redundant to check bounds in
isMoveValid() method, since it has already been done in the parent
if-condition. So that has been removed.