client/pieces: implement notifying square of King's influence
Like with Knight, all squares within King's range from its position should
be notified, regardless of occupational status. Therefore isInBounds()
is called for ensuring that the board position exists before we add the
piece to the square on that position.

Due to the manner in which the changes are made, it is redundant to
check the bounds of a square inside isMoveValid(), therefore it now only
checks if the square is empty or if it contains an opponent piece.