client: remove pieces from squares that no longer influence the square
After a move, some pieces may not influence some squares any more. These
pieces should be removed from the square's m_pieces vector.

This is accomplished by maintaining a vector (in addition to m_moves) in
each piece that stores all the squares it influences. m_moves stores
only those that the piece can be moved to.
This new vector is changed whenever a new move has been made on any
square that it previously influenced.