client: trim Game class manage only one player
The Game class earlier managed both players and thus the entire game.
Since this is an online, multiplayer game, the Game class should deal
with only its own client. As a result, black and white instances of
Player class are removed. The Game class now just holds the colour of
its client.

In the process of doing this, Game constructor was also modified. There
were other smaller changes, such as replacing sf::Color with
Piece::Color in some places.