client,server: notify client when game can begin
When the second player joins the server, the client is notified that the
game may begin now. Until this happens, the client may not allow the
user to play any moves. Presently this "not allowing" takes the form of
the client being non-responsive to the user's actions, but later
a message should be put out.

Also new message format replaces the old one in this commit.